WL#11651 \show thread and \show threads
This commit adds two new reports which can be invoked using \show and
\watch commands:
- thread - Provides various information regarding the specified thread,
by default lists information on the thread used by the current session.
- threads - Lists threads that belong to the user who owns the current
session.

For more information on these reports run:
\show thread --help
\show threads --help

Additional changes:

1. Short options used both when running the Shell or executing a report
can be chained, i.e.:
  mysqlsh -AEifull
is equivalent to:
  mysqlsh -A -E -i full
If any of the chained options is expecting a value, the remaining
characters are treated as this value.

2. When registering a report using the shell.registerReport() method, it
is now possible to add example usage of the new report. This can be done
using new 'examples' key in the 'description' dictionary. The associated
value should be a list of dictionaries, each of which contains following
keys:
- description (string, required) - Description text of the example.
- args (list of strings, optional) - List of the arguments used in the
  example.
- options (dictionary of strings, optional) - Options used in the
  example with their corresponding values.

3. When registering a report using the shell.registerReport() method, it
is now possible to specify that a string option used by the report
accepts an empty value. This can be done using a new key named 'empty'
holding a Boolean value, specified in a dictionary describing an option.
This allows to register a report which has an option which optionally
expect a string value, i.e.:
  \show my-report --opt - the string option named 'opt' will be empty
  \show my-report --opt value - the string option will contain 'value'

Change-Id: I07764b44aa198bafa89d5550abb6ab266e55320b