Fixing duplicate entries in dir() result for Python
When overloading functions with expose(), separate entries with the
same name are registered on the function lists in the Cpp_object_bridge.

The dir() function in python calls get_members which returns the list of
all the entries registered, including the duplicate ones, resulting in
duplicate entries for the result of the dir() call.

The problem was not detected by the Unit Tests because of a weak
member validation logic, all the UTs have been updated to properly
validate the returned members on each object vs a list of expected
members, generating an error if:

- Unexpected members are found
- The member list is missing expected members

Change-Id: Ic44c01360a356f1a230f98769105f113688f9a29