WL#12776: AdminAPI: Refresh internal accounts passwords
This patch introduces a new command for the AdminAPI on the cluster
object: <Cluster>.resetRecoveryAccountsPassword.

This command allows users to reset the passwords of all internal
recovery accounts used by the cluster, setting a new random password
for each account. By default the operation will abort if any member is
not reachable (OFFLINE, MISSING or RECOVERING) unless a force option
is set.

New test cases were added to test the new functionality.

Change-Id: Id73288c3673dc21910b83bead7869e3eb72f8b07