BUG#30156304 PYTHON LAMBDA REGISTERED AS AN EXTENSION OBJECT MEMBER CAUSES SHELL TO CRASH
When Shell shuts down, the JS context is first to be destroyed, which
in turn causes all extension objects and their members to be destroyed
as well. If a member is a Python lambda, Shell holds the sole reference
to this Python object, and it is immediately released.

Shell was crashing, because Python object associated with the lambda was
released without properly entering the Python context, this is now
fixed.

Change-Id: I3b0e993202f3bbbf76922319bf755502cb15ff7d