BUG#30138755 MEMORY LEAK WHEN EXECUTING QUERY IN A LOOP (PYTHON)
When Shell was running in interactive mode and user executed Python
code which contained a loop where each iteration produced some temporary
value(s), Shell would leak memory and become unusable when loop was
finished.

Shell was storing each value produced when a Python code was
interactively executed, causing long-running loops to continue to
allocate memory, while only the last one was actually used to present
the output to the user. The remaining values were stored, and could be
returned to the user if another code which did not produce a value was
executed, causing confusion.

This behaviour is now fixed, Shell will store only one Python value at a
time, preventing the unnecessary memory allocations.

Change-Id: I9b56e0c102bdd9965a54ad79e7716e5d235fd085