BUG#30171090 : configureInstance does not set binlog_checksum = none
The configureInstance() operation was not setting the binlog_checksum option
with the required value (NONE) in the option file (my.cnf) for instances that
do not support SET PERSIST (e.g., 5.7 servers), when the option file path was
not provided as input parameter but instead specified though the operation
wizard (interactive mode).

This happened because the operation was only requiring read-only variables
to be persisted in the option file, when SET PERSIST is not supported,
instead of requiring all needed variables to be set (persisted) in the option
file (even dynamic ones).

This patch fixes the issues by requiring all variables, independently if
they are read-only or not, to be persisted in the option file if the SET
PERSIST feature is not supported by the instance (e.g., 5.7 servers).

A test was added for regression.

Change-Id: I5d05afb0809b24ac18a62d7e78fe438c7a894b90