BUG#29954085 : STUCK TRYING TO USE CLONE RECOVERY WHEN CLUSTER CREATED USING OLD SHELL
<Cluster>.addInstance() did not ensure the clone plugin was
installed/loaded on all cluster members, when available and not
disabled.

For that reason, whenever a cluster was created using an older Shell
version, on a target MySQL instance supporting clone, the instance would
not have the clone plugin installed so any .addInstance() call that
used clone would fail. The same issue would happen if an instance was
added to a cluster of just 1 member using the incremental recovery type
and afterwards the seed instance removed. That results in all cluster
members not having the clone plugin installed and thereby any instance
added using the clone recovery method would fail.

This patch fixes the issues by ensuring that the clone plugin is
installed on all cluster members (if available and not disabled) at
cluster creation time and also whenever an instance is added to a cluster.

Added tests for regression and updated existing tests to also validate
this requirement.

Change-Id: Iff97de513ec1129a1cf8b9ad4d2af05061b752f2