BUG#30174191 : auto_increment policy does not not follow correct behavior
The <cluster>.rejoinInstance() operation was not setting the auto increment
values (policy) defined for the AdminAPI, leading to the use of the
default Group Replication behavior if the instance configurations were not
properly persisted (using dba.configureLocalInstance()) for 5.7 servers.

This patch fixes this issue by updating the auto-increment settings of the
target instance for the <cluster>.rejoinInstance() operation.

A test was added for regression.

Change-Id: Ic7ca9f7379848d5c2a33e1ea2fa8506a1af8e38d