BUG#29953812 : ADD_INSTANCE() PICKY ABOUT GTID_EXECUTED, REJOIN_INSTANCE() NOT: DATA NOT COPIED
<Cluster>.rejoinInstance() does not check the GTID consistency of the
target instance in regard to the cluster. An instance with errant
transactions is allowed to join the cluster possibly provoking diverged
data sets.

On the other hand, the command also does not check if the binary logs
have been purged from all cluster members resulting in a failed attempt
to rejoin the instance to it as distributed recovery fails.

Finally, a check for an empty GTID set was missing. An instance rejoining
a cluster with an empty GTID set results in a failed attempt to rejoin.

This bug improves the GTID consistency checks in
<Cluster>.rejoinInstance() by checking for irrecoverable or diverged
data-sets and also for empty GTID sets, to promptly terminate the command
with a descriptive error. The user is also informed with the list of
errant transactions, possible outcomes and solutions.

Added tests to verify the new functionality added.

Change-Id: I4d27f1ea89c6e17ed903aecc60b0c8dd040ad1f9