WL#13462, WL#13466 listRouters() and removeRouterMetadata()
This patch introduces the following function:

Cluster.listRouters([options])

The function will list the routers associated to the Cluster.

The option "onlyUpgradeRequired" will make the function to only
list the routers that are not compatible with the supported
metadata.

WL#13466 AdminAPI: Unregister Routers (that do not support latest metadata)

Remove router instances from the metadata schema.

Change-Id: Idc1b92c299db3dd87f051011a88822411e0d641c