BUG#30333801 WINDOWS MSI DOES NOT REMOVE THE LIB DIR IN CUSTOM PATH
When Shell was installed on Windows in a custom location, running the
MSI to remove the installed application left some files and folders in
the installation folder.

If Shell is installed in a location, where the user has full write
permissions, the Python library used by Shell will create compiled files
and additional folders whenever a Python module is imported. Since
Windows installer removes only those files which were created during the
installation process, any files created in that folder after the
installation are not removed.

This issue is now fixed: when Shell is uninstalled on Windows, installer
will remove all additional files created by the Python library.

Change-Id: I9946bb242e9315fecc82f6bf6def6eb96bcd4bbe