WL#11033 AdminAPI: upgrade metadata schema
This patch introduces the required engine to perform cluster metadata
upgrades. The Dba.upgradeMetadata() function is added to detect and perform
metadata upgrades up to the current version supported by the shell.

The engine supports a basic upgrade from metadata version 1.0.0 to
version 1.0.1 with the only purpose to create the schema_version table.

In addition, the AdminAPI functions now will also evaluate the metadata
state and be enabled on different criteria. The metadata state is the
result of a semantic comparison between the version of the installed
metadata and the version of the supported metadata.

Available States:

- Nonexisting: when no metadata is installed.
- Compatible: when the installed and supported version match on the major
  version.
- Incompatible: when the installed and supported metadata have
  differences in the major version.
- Upgrading: When the metadata is in an upgrade progress.
- Failed Upgrade: When the metadata upgrade process got interrupted.

The Admin API functions will have an additional precondition check that
will enable/disable them based on the metadata status.

- All the functions are available if the metadata is in a Compatible state.
- dba.getCluster: unavailable during an upgrade process. Available on any
  other state but if in an Incompatible state a warning will be shown
  indicating the state and no changes will be allowed on the cluster.
- Cluster read only operations will be available as long as the metadata is
  not in an upgrade progress.
- Write operation will be unavailable if the metadata is either in an
  upgrade process or incompatible.

The following functions do not have precondition checks and so are not
affected:
- Sandbox operations
- dropMetadata
- configureInstance
- configureLocalInstance
- checkInstanceConfiguration
- checkInstanceState

Change-Id: I9aa94e735f4b0f32a73f18ab8d73f3b88492627e