WL#13119 Enhancements to the require() function for JavaScript
The JavaScript function:
  any require(String module_name_or_path)
has been modified to allow loading local modules, relative to the file
or module which is currently being executed by the Shell. Loaded modules
have access to more information via new variables (__filename - full
file path of the module, __dirname - full path to the containing
directory), and are able to modify the type of exported functionality via
the module variable.

A module can be a single JavaScript file or a JavaScript plugin folder
containing the init.js file.

Change-Id: I8e21aa253de5cf1f668dec7cc0c12bbda7f153a6