BUG#30339460 : Access denied for dba.configureInstance()
Some privileges are missing for the configure instance operations
in order to persist any required configuration changes for 8.0 MySQL
servers. In particular, an "Access Denied" error is issued indicating
the SYSTEM_VARIABLES_ADMIN and PERSIST_RO_VARIABLES_ADMIN privileges
are required.

This issue occurs because the AdminAPI is missing to validate and to add to
the InnoDB Cluster Admin user the above privileges, for 8.0 MySQL servers.

This patch fixes the issues by adding SYSTEM_VARIABLES_ADMIN and
PERSIST_RO_VARIABLES_ADMIN privileges to the list of required privileges
for the InnoDB Cluster Admin user, properly validating the need for those
privileges and adding them to the Admin user created by
dba.configureInstance() for 8.0 MySQL servers.

Tests were appropriately updated and new added for regression.

Change-Id: Ic41499e24043e0b54e7711700d7f0d68a265e034