BUG#30029568 PYTHON IN INTERACTIVE MODE IS UNABLE TO EXECUTE MULTIPLE STATEMENTS AT ONCE
When shell was running in an interactive mode, it was not possible to
execute multiple statements at once in Python mode.

This is now fixed, if user tries to execute multiple statements (i.e.
several function definitions) or uses \edit command to write i.e.
several assignment expressions, all of them are going to be executed
correctly.

Change-Id: I95a3c4bfce761c8188e0b72960374bcdcc9dd129