WL#13236 Shell integration with InnoDB ReplicaSet
This worklog adds new command line option, new method and modifies the
behaviour of some of the existing command line options, in order to make
working with ReplicaSets using Shell easier.

New command line option
 --replicaset
Ensures that the target server is a member of a ReplicaSet and if so,
sets the rs global variable with the ReplicaSet the target instance
belongs to.

New method
 Session shell.connectToPrimary([ConnectionData connectionData][, String password]) (JS)
 Session shell.connect_to_primary([ConnectionData connectionData][, str password]) (PY)
Ensures that the target server is a member of a cluster or a ReplicaSet
and if it is not a PRIMARY, finds the cluster's PRIMARY and connects to
it. Sets the global session object to the established session, returns
that object. If connectionData is not given, uses the
active shell session, if there is none, raises an exception.

Modified command line options
 --redirect-primary
 --redirect-secondary
Both options can now be used with ReplicaSets.

Change-Id: I0b8ebf179997a8f1bdb733bef514235f40ada554