WL#13540 : Locking mechanism for ReplicaSet operations
This patch introduce a locking mechanism to manage the concurrent execution
of InnoDB Replica Set operations. A lock error is now issued when an
operation is not allowed to be executed because another one is still in
progress. The following AdminAPI operations are covered by this patch:
- Operations requiring exclusive access:
    - dba.upgradeMetadata();
    - dba.createReplicaSet();
    - dba.configureReplicaSetInstance();
    - <ReplicaSet>.forcePrimaryInstance();
    - <ReplicaSet>.setPrimaryInstance();
- Operation with shared access on the primary but exclusive on the target
  instance:
    - <ReplicaSet>.addInstance();
    - <ReplicaSet>.rejoinInstance();
    - <ReplicaSet>.removeInstance();
    - <ReplicaSet>.removeRouterMetadata();
- Read operations without locks:
    - dba.getReplicaSet();
    - <ReplicaSet>.status();

New tests were added for regression.

Change-Id: I57509415335b59e2b018fd0007df9bb58cf23003