Bug#30533318 UNRELIABLE/MISLEADING CONNECTIONS WITH MYSQL_UNIX_PORT
Connecting to socket path specified through MYSQL_UNIX_PORT does
not work if the environment variable was set within the same process
that opens the connection, because libmysqlclient caches it during
initisalization. OTOH the shell would assume it got connected to that
socket, even if it didn't.

This fixes the session connect code so that the actual socket path is
displayed to the user.

Also, don't require the host param when connecting using a dictionary of
connection options.

Change-Id: I2147da29877a61548835596323c6d5ff40c2dc78