WL#13237 BUG#30360185 BUG#30466618 BUG#30273760 BUG#30633382 BUG#30618211 BUG#30630111 BUG#30628746 BUG#30624443 BUG#30629831 BUG#30630023
WL#13237: Clone support for Async Replica Sets

This patch introduces support for clone in InnoDB ReplicaSets.

The AdminAPI for ReplicaSet was extended to to support using the clone plugin,
when available, to provision instances when:

  - Adding a new member to a ReplicaSet
  - Rejoining a member to a ReplicaSet

On top of that, automatic detection of the most appropriate recovery method
was added to guide the users, and also monitoring capabilities.

The following high-level features/changes are introduces:

<ReplicaSet>.addInstance()/.rejoinInstance():
  - extended with a new option: recoveryMethod to allow the user to explicitly
    select which recovery method must be used: auto (default), clone or
    incremental
  - extended with checks to determine which type of recovery must be used and
    prompt the user accordingly (when auto is selected).
  - extended with a new option: waitRecovery, to indicate the level of
    verbosity of the recovery process.
  - extended with a new option: cloneDonor, to select a donor from the
    ReplicaSet, when clone recovery is used.
  - extended with a new option: timeout, to select the maximum number of
    seconds to wait during the synchronization period.

Support for canceling the recovery process at any time using ctrl-c was also
added, ensuring the operation and changes are fully reverted.

Added new tests to cover the new features and updated existing ones.

BUG#30360185: ASYNCREPLICASET.ADDINSTANCE() - RECOVERYMETHOD:AUTO IS NOT
              REALLY AUTO

The initial implementation of ReplicaSet included the optional recoveryMethod
option in .addInstance(), however, it only accepted the values 'auto' and
'incremental' and it did do any automatic detection.
This patch fixes the bug by including support for 'clone' as another available
recovery method and also by including the automatic detection and guidance for
the most appropriate recovery method.

BUG#30466618: REPLICASET.REJOIN_INSTANCE REPORTS SUCCESS BUT INSTANCE REMAINS
              IN ERROR STATE

Attempts to rejoin an instance that got its GTID diverged in regards
to the ReplicaSet, using .rejoinInstance(), would not fail but would also not
succeed because the instance would go into ERROR state.

This patch fixes the issue by:

  - Detecting erroneous transactions when adding or rejoining instances to a
    ReplicaSet.
  - On such scenarios, if clone is available prompt the user to use it,
    otherwise fail with an error.

BUG#30273760: SHELL CRASHES ON SIGINT WHEN WAITING FOR SERVER TO RESTART

Whenever ctrl-c was used to abort waiting for a server restart (when clone is
used as recovery method), the Shell would crash.

This issue applies for both InnoDB Cluster and InnoDB ReplicaSet. Considering
that there's no reason to allow aborting the processing (or sending it to
background) when the target instance is restarting, the ability for such
operation was disabled. With this patch, when ctrl-c is pressed during the
restart phase of the recovery it's ignored.

BUG#30633382: REPLICASET - CAN'T DO CLONE RECOVER USING CLUSTERADMIN ACCOUNT

clusterAdmin account was missing privs for clone

also added a bunch of new tests

BUG#30618211: <REPLICASET> REJOININSTANCE WITH RECOVERYMETHOD AS CLONE MAKE
              MYSQLSH TO EXIT

<ReplicaSet>.rejoinInstance() using clone as the recoveryMethod caused
the Shell to exit and not terminate the execution of the operation.

This was caused by a segmentation fault triggered by a failed assertion.
The failed assertion happened on the method to sync transactions, which
requires a connection to the primary member of the ReplicaSet that it
was previously cleared/closed in the caller.

This patch fixes the issue by removing the uncessary code added to reset
the primary instance variable and acquiring a new connection to it in
rejoinInstance().

Added a new test to verify the usage of clone in rejoinInstance().

BUG#30630111: .ADD_INSTANCE(CLONE): "WAITRECOVERY": 0 -> IT MUST BE AN INTEGER
              IN THE RANGE

Shell was printing 0-3 as valid range when the right range is 1-3.
Fixed printed range and added test case.

BUG#30628746: ADD_INSTANCE: CLONEDONOR FAILS, USER DOES NOT EXIST

When using clone recovery to add or rejoin an instance to a ReplicaSet
and specifying a donor that is processing transactions, the command could
fail with an error indicating the clone recovery account didn't exist.

This happens when the donor is processing transactions and hasn't
processed yet the ones referring to the clone user handling (create +
grant), resulting in a failure of the clone command since the clone user
does not exist in the donor yet.

This patch fixes the issue by adding a sync between the donor and the
primary instances before starting the clone itself, to ensure that the
clone user handling transactions were already applied.

Added regresssion tests and new tests for rejoinInstance() using the
cloneDonor option.

BUG#30624443: RS.ADD_INSTANCE() HELP + CLONE: PRECONDITIONS SECTION

The online helper of <ReplicaSet>.addInstance() includes a section about
pre-requisites to run the command which was outdated since the
introduction of clone as a possible recovery method.

This patch fixes the issue by improving and correcting the
pre-requisites section ensuring the information differentiates the
scenarios on which incremental recovery is used VS clone.

On top of that, the <ReplicaSet>.rejoinInstance() was missing a
pre-requisites section. This patch also adds the missing section in that
function.

Updated tests to meet the changes.

BUG#30629831: RS.ADD_INSTANCE(CLONE): IPV6 ADDRESS REPORTED BY PRIMARY VS.
              CLONE PLUGIN IPV4

- Validate cloneDonor against IPv6

- Check if target hostname/report_host is IPv6 when verifying
compatibility for clone

- Added handling for case where no members are compatible to be used as
clone donors, including PRIMARY

- Full tests for IPv6 + replicasets

- Added addInstance and rejoinInstance with clone tests to all simple_*
replicaset tests

- Fixed unreported bug where mysqlX IPv6 addresses were stored
incorrectly (::1:33060 vs [::1]:33060)

- Fixed cluster adopt with IPv6 addresses

- Added clone test in all simple_* tests for replicasets

- Added make_host_and_port() and split_host_and_port() to normalize
host:port and IPv6 handling

BUG#30630023: RS.ADD_INSTANCE(CLONE): ONLINE HELP SHOULD BE IMPROVED

The online helper of <ReplicaSet>.addInstance()/.rejoinInstance() was
incomplete in comparison to its counterpart in InnoDB Cluster.
Specifically, it was missing information about the options related to
instance provisioning.

This patch fixes that by adding more information in the online helper
of those functions about the options:

- recoveryMethod
- waitRecovery
- cloneDonor

A description of each option was added, with the explanation of every
accepted value, and possible limitations.

Updated tests to meet the changes.

Change-Id: I56b34acc2f968a8bcb5e6059aa7b633138b3e8d4