WL#13019 DevAPI: Add schema validation to create collection
This WL introduces support for JSON schema validation for a collection.
While being able to insert any type of document into a collection is useful,
being able to specify the proper structure of a collection is also very useful.
It has the benefit of requiring documents to have a certain structure before
they are allowed to be inserted or updated.

Functionality is available via methods of Scheme class:
- createCollection method has been extended to accept optional options
  dictionary with validation specification e.g (Javascript):

  var coll = schema.createCollection("mycollection", {
    validation: {
      level: "off|strict",
      schema: "object"
    }
  })

- modifyCollection method has been added, witch allows to modify validation
  specification (only the changed part of validation object is necessary) e.g.:

  schema.modifyCollection("mycollection", {
    validation: { level: "off|strict"} })

Additionality createCollection method accepts 'reuseExistingObject' Boolean
option that when defined as 'true' will cause call to succeed even when
collection with the same name already exists.

Change-Id: I1093eee8f79c09735a5957ffe507f2784946ae3a