BUG#30391076 CALLING ADD() ON A COLLECTION IN A LOOP ONLY ADDS A SINGLE DOCUMENT
The behaviour described in the bug report is not a bug, the DevAPI CRUD
operations only perform a modification to the database when the
execute() method is called.

If Shell is running in interactive mode, it performs additional
handling of the values returned by the executed code:
 - DB results are automatically printed out to the console,
 - DevAPI CRUD operations (CollectionAdd, CollectionFind,
   CollectionModify and CollectionRemove) are automatically executed.
Running a loop in the interactive mode has a side effect in both
scripting modes: the value of the last statement of the last iteration
is returned (i.e. as if it was executed by hand). In combination with
the automatic DevAPI handling described above, this results in the
reported behaviour: only the last document is added to the database.

Adding documents in a loop and calling execute() in each iteration:
  for x in data: collection.add(x).execute()
will correctly add all documents, however Shell will print information
that only one record was added. This is caused by the fact, that the
execute() method returns a Result object, and as explained above, the
last Result instance is returned from the loop and automatically printed
out.

The easiest way to add multiple documents is simply to use the array
itself:
  collection.add(data).execute()
or in interactive mode:
  collection.add(data)

Since the docs suggest that the add() method modifies the database,
this bug was treated as a documentation issue.

The documentation of the DevAPI CRUD operations has been updated in
order to emphasize that no modifications to the database are done until
the execute() method of the corresponding object instance is called.

Additionally:
 - the related documentation has been cleaned up,
 - missing method links were fixed,
 - deprecated methods are no longer explicitly mentioned by the docs,
 - documentation of Table and related CRUD operations has been reviewed
   and fixed.

Change-Id: I43430e226f49854826b37993e5d628d5f4e76042