BUG#29705983 : cluster.status error: get_uint(24): field value out of range
Somtimes the status() operation could report an error "get_uint(24): field
value out of the allowed range" because it was always expecting a positive
values for some fields that could in fact have negative values. This
can happen when the clocks of different members of the clusters are skewed.

This issue was fixed by allowing negative values to be returned for those
time fields. Although those values can be considered "invalid", they should
be ignored by the operation and still produce the result (instead of issuing
an error).

NOTE: two servers on different machines with skewed clocks are required
      to reproduce this issue.
Change-Id: I3a034b6a3e491fbb650fda6865fb4f4686377e32