BUG#30580393 : Remove_instance(user@host:port) poor Error Message
The <Cluster>.removeInstance() operations was issuing a misleading error
message when the instance was unreacheable, indicating that it does not
belong to the cluster when an alternative (valid) host/IP is used.

This error can be considered incorrect since in reality the instance
belongs to the replicaset but the host/IP registered in the metadata does
not match the one used as parameter for the operation. Moreover, this
error message hides eventual credential issues that should be displayed
to the user (e.g., in case a wrong login or password was used).

Thus, an error indicating that the instance is unreachable should preferred
in this situation, being more accurate and encouraging users to only remove
available instances (ONLINE).

This patch fixes the error messages displayed when the instance is not
reachable (using the specified connection options), providing more accurate
information to users.

New test cases added for regression.

Change-Id: I4ee38714b8debd18f6d0c8f537182137043e8edb