Bug#30501628,Bug#30625424 removeInstance() bug fixes
removeInstance() was not correctly handling some cases:

- if the instance is ONLINE, it should be removable by any address that
can reach the instance, as long as the report_host and the address in
the MD match

Bug #30501628	REMOVEINSTANCE, SETPRIMARY ETC SHOULD WORK WITH ANY FORM OF ADDRESS
- if report_host does NOT match what's in the MD but the address
given in the command does match it, then it should be possible to remove
the instance.

- similarly, if the instance is unreachable, it should be possible to
remove it if the given address matches what's in the MD

- case where instance was OFFLINE but reachable (GR stopped but server up),
was not handled. Instance won't sync, so force should be required.
Also, sync shouldn't be attempted if force=true

Bug #30625424	REMOVEINSTANCE() FORCE:TRUE SAYS INSTANCE REMOVED, BUT NOT REALLY
- if the instance was OFFLINE but reachable, removing the instance
through an address that's not what in the MD would make the command
succeed but do nothing.

- Additional missing tests were added for removeInstance().
Same for setPrimaryInstance(), which was working correctly.

Change-Id: I19c343ca9c0f6dde71f8c9f94fe4a7d5991fe973