BUG#30406283 ILL-FORMED UTF8 IN PROMPT FILE CAUSE ERROR MESSAGE IN PROMPT
If a prompt file contained an ill-formed UTF-8, the prompt text contained
an error message instead of a sensible default.

This error is now fixed, the prompt file is validated before it is
loaded, if it contains an ill-formed UTF-8 sequence, an error is
displayed and the default prompt is used instead.

Change-Id: Iad29dddf48d7277ea6e2a6f3a0ee3179ea82b07e