BUG#30594844 : remove_router_metadata() gets primary wrong
The removeRouterMetadata() operation was issuing an error updating the
metadata when executed on a ReplicaSet object created with a session to
a secondary. This error was issued because the operation was trying to
updated the Metadata on the secondary (read-only) instead of doing it
on the primary.

This patch fix this issue by obtaining the primary of the replicaset to
update the metadata on it, instead of using the current MySQL shell session
as target. The operation was working as expected in the case of clusters
(no change needed).

New tests added for regression (for both ReplicaSet and Cluster).

Change-Id: I8d1e12da7dd772129f526c54c8a83849aea7ca25