Fix for infinite loop in one of the tests on Windows
The test Admin_api_scripted/Auto_script_js.run_and_check/cluster_addinstance
was not terminated on Windows due to a bug in Uri_parser: it failed to
parse an invalid URI `[[::1]]:26061`, disallowed characters were not
checked at the beginning of the IPv6 address.

This URI was constructed due to another bug, Net::is_ipv6() method on
Windows returned true for addresses in form `[IPv6]`, this is also
fixed.

Change-Id: I09b74a4f378787d6b5407a2f33c18dee59800e7b