BUG#30632029 : RS.ADD_INSTANCE(CLONE): ERROR SETTING UP REPLICATION ON NEW SANDBOX...
When the target MySQL server does not have --skip-slave-start enabled or
is a fast system, adding an instance to a ReplicaSet using clone as the
recovery method and a secondary as donor could fail with MySQL Error 3081.

This issue happens because clone copies all tables and settings, including
the replication settings. If the donor is a slave, its replication settings
are copied over and since MySQL starts replication by default if
--skip-slave-start is not enabled, any CHANGE MASTER query fails with
error 3081.

This patch fixes the issue by ensuring that the slave is stopped and the
replication channels are reset on the target instance before the new
CHANGE MASTER is performed.

Change-Id: Ib94078da2a8f87e663ffd2ebdf1c3cf07f9c50b4