BUG#30657204 : configureLocalInstance() fails with key not found (LogicError)
The configureLocalInstance() operation fails with "key not found
(LogicError)" when executed on an instance (not a sandbox) where it
does not have access to the (found or prompted) my.cnf file and
the operation request an output configuration file to be specified.

This issue occurred because the code was internally trying to modify a
Config handler object that did not exist to set the new output
configuration file path.

This patch fixes this issues by removing the harming code, since it is not
required. The update of the internal update of the Config handler object
with the new output configuration file path is already properly performed
later (by the caller function).

Test was added for regression.

Change-Id: I69dd68431310e6b5e4d3a122f4e774a7b6a9ff3d