BUG#30281908 : MYSQL SHELL CRASHED WHILE ADDING A INSTANCE TO INNODB CLUSTER
Adding an instance that does not support RESTART to a cluster, using
clone as the recovery method caused the shell to crash.

This was caused by a segmentation fault in the post-clone restart
handling that did not cover a possible timeout when waiting for the
instance to start after the restart performed by clone.

This patch fixes the issue by implementing the logic to handle this
scenario and also to print informative messages to the user indicating
that <Cluster>.rescan() must be used to ensure the instance is added to
the Metadata. This is the right approach considering that Group
Replication handles the clone and after the instance is restarted it
will automatically join the group.

Added regression tests.

Disabled _norecord tests for platforms on which the hostname resolves to
127.0.1.1.

Change-Id: I29422714bd9cfadc3ed80a58f9f99df3b5684b66