Bug#30609075 INNODB CLUSTER TO STANDALONE SLAVE REPLICATION FAILED DURING CLONE
This bug happens when there's an async slave of a
cluster member and a replication filter was setup
to filter out the MD schema.

The issue happens when an instance is added to the
cluster using CLONE, because that operation will
try to GRANT a privilege on an account that was
created during createCluster(). The account creation
is filtered out by the slave because the MD schema
was the default DB at the time. So trying to GRANT
a privilege on that account fails and breaks replication.

createCluster() changes the default DB to the MD schema
while executing DDL. The slave will filter out anything
at all that's executed while the MD schema is the default,
including unrelated CREATE USER stmts.

So if addInstance() were executed in the same session
as the createCluster(), the error doesn't happen because
the GRANT is executed while the MD schema is still the
default. But if it's executed in a new session, it will
fail.

The solution is to always change the default DB from the
MD schema to something else (USE mysql), right after the
MD schema is installed or upgraded. These are the only
operations that change the default DB in the AdminAPI.

Change-Id: Ia79f4ccf5a50fa8a47abc804f04a0b7624ea9bc2