BUG#30394258 BUG#30401048 AdminAPI does not handle high transaction consistency levels of Group Replication
The AdminAPI ensures that all members of a cluster have the same
consistency level as it was set on cluster creation time. However, when
high and non-default consistency levels are chosen for the cluster, adding
instances to it results in error 3796 which indicates that
group_replication_consistency cannot be used on the target instance. This
happens because the consistency values of "BEFORE", "AFTER" and
"BEFORE_AND_AFTER" cannot be used in members that are RECOVERING and
several transactions happen meanwhile the instance is in the RECOVERING
phase (BUG#30394258)

Other AdminAPI commands will result in the same error for the same
scenario (high global consistency levels) whenever at least one member
of the cluster is RECOVERING. For example, dba.getCluster() (BUG#30401048).

This patch fixes the issue by ensuring that all sessions used by the
AdminAPI use the consistency level of "EVENTUAL" when the consistency
setting is available, i.e. MySQL version >= 8.0.14.

Added tests to cover the different possible scenarios that cause this
issue.

Change-Id: I9a02ba1e01e61c340ebef586c8b6f54352688528