BUG#29654346 : CLUSTER.OPTIONS() IS MISSING AUTOREJOINTRIES
In 8.0.16, a new option was added to define the number of times an instance
will attempt to rejoin the cluster after being expelled: 'autoRejoinTries'.
The option is a valid cluster setting, controlable through the AdminAPI
like many other options.

However, the option is not listed in <Cluster.>options().

This patch fixes the issue by ensuring the option is listed in the
output of <Cluster>.option() accordingly.

Updated tests to meet the changes.

Change-Id: I826f21c279edc7091a7ec0e02e10383764c28a9d