WL#13536 BUG#30640058 BUG#30676863 BUG#30645140 BUG#30648813 AdminAPI: Setup cluster admin and Router accounts
This patch introduces two new command for the AdminAPI on both the
cluster and replicaset objects: setupAdminAccount and
setupRouterAccount.

The setupAdminAccount allows you to create a new MySQL account (or upgrade
an existing MySQL account) with the necessary privileges to admin an
InnoDB cluster/replicaset.

The setupRouterAccount allows you to create a new MySQL
account (or upgrade an existing MySQL account) with the necessary
privileges to use with MySQL Router.

New test cases were added to test the new commands and their
arguments.

BUG#30640058: shell's setuprouteraccout() gives unnecessary "with grant option"

This bug was fixed by removing the WITH GRANT OPTION that was given with the
privileges granted to new accounts/existing accounts by the setupRouterAccount
operation. Tests were added to verify the fix and to avoid regressions.

BUG#30676863: shell's setupadminaccount() gives permissions different than
listed in WL.

This bug was fixed by adding the CLONE_ADMIN and EXECUTE privileges to the list
of privileges granted by the setupAdminAccount when primary server version is >=
8.0.17. Tests were added to verify the fix and to avoid regressions.

BUG#30645140: shell's new commands allow user@host1@host2

This bug was fixed by adding a new validation that prevents users from using
the `@` character on the username (unless the username is surrounded by quotes).
Tests were added to verify the fix and to avoid regressions.

BUG#30648813: shell's new commands with user = '@%' not consistent with user = ''

This bug was fixed by adding a new validation that prevents users from
specifying empty usernames on the account parameter.
Tests were added to verify the fix and to avoid regressions.

    dba_cluster_help.js
    setup_admin_account.js
    setup_router_account.js
    bug26737608_cluster_admin_should_have_less_grants.js
    bug27324699_createcluster_throws_segmentation_fault_interactive.js
    dba_cluster_help.js
    dba_upgrade_metadata.js
    dba_upgrade_metadata_dbug.js
    setup_admin_account.js
    setup_router_account.js
    setup_admin_account.js
    setup_router_account.js
    rs_help.js
    setup_admin_account.js
    setup_router_account.js
    dba_cluster_help.py
    dba_cluster_help.py
    dba_upgrade_metadata_norecord.py
    dba_cluster_interactive.js
    dba_cluster_no_interactive.js
    dba_check_instance_configuration_session.js

Change-Id: I0883bb74012be7cfe86e4da38a57011ae2d23f8d

    dba_cluster_help.js
    setup_admin_account.js
    setup_router_account.js
    bug26737608_cluster_admin_should_have_less_grants.js
    dba_cluster_help.js
    dba_upgrade_metadata.js
    dba_upgrade_metadata_dbug.js
    setup_admin_account.js
    setup_router_account.js
    setup_admin_account.js
    setup_router_account.js
    rs_help.js
    setup_admin_account.js
    setup_router_account.js
    dba_cluster_help.py
    dba_cluster_help.py
    dba_upgrade_metadata_norecord.py
    dba_cluster_interactive.js
    dba_cluster_no_interactive.js
    dba_check_instance_configuration_session.js

Change-Id: I03f18afbbe9cc0c0649a9db234436e51ebdf6f72